Moved SDK layer out of ceph directory
