+++ cephV2/ceph/osd.py
@@ -4,7 +4,7 @@ This module is to contain CLI wrapper for all `ceph osd ` commands
 and modules for sub-commands of `ceph osd ` would use this module.
 """
 
-from ceph.cephV2.ceph import Ceph
+from cephV2.ceph import Ceph
 from utility.log import Log
 
 log = Log(__name__)
