+++ cephV2/rbd/feature.py
@@ -1,4 +1,4 @@
-from ceph.rbd.rbd import Rbd
+from cephV2.rbd.rbd import Rbd
 from utility.log import Log
 
 log = Log(__name__)
