+++ cephV2/rbd/mirror/mirror.py
@@ -1,10 +1,10 @@
-from ceph.rbd.mirror.pool import Pool
+from cephV2.rbd.rbd import Rbd
 from utility.log import Log
 
 log = Log(__name__)
 
 
-class Peer(Pool):
+class Mirror(Rbd):
     def __init__(self, nodes):
         self.nodes = nodes
-        super(Peer, self).__init__(nodes=nodes)
+        super(Mirror, self).__init__(nodes=nodes)
