+++ cephV2/rbd/rbd.py
@@ -1,4 +1,4 @@
-from ceph.cli import CephCLI
+from cephV2.cli import CephCLI
 from utility.log import Log
 
 log = Log(__name__)
